move sdpconsistency to es6 style
